Add support for SQL Azure database edition options
Fixes #2969
Fixes #15004

+++ test/EFCore.Cosmos.FunctionalTests/TestUtilities/CosmosTestStore.cs
@@ -21,6 +21,8 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.TestUtilities
         private readonly TestStoreContext _storeContext;
         private readonly string _dataFilePath;
         private readonly Action<CosmosDbContextOptionsBuilder> _configureCosmos;
+        private bool _initialized;
+
         private static readonly Guid _runId = Guid.NewGuid();
 
         public static CosmosTestStore Create(string name, Action<CosmosDbContextOptionsBuilder> extensionConfiguration = null)
@@ -72,6 +74,7 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.TestUtilities
 
         protected override void Initialize(Func<DbContext> createContext, Action<DbContext> seed, Action<DbContext> clean)
         {
+            _initialized = true;
             if (_dataFilePath == null)
             {
                 base.Initialize(createContext ?? (() => _storeContext), seed, clean);
@@ -219,7 +222,8 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.TestUtilities
 
         public override async Task DisposeAsync()
         {
-            if (_dataFilePath == null)
+            if (_initialized
+                && _dataFilePath == null)
             {
                 await _storeContext.Database.EnsureDeletedAsync();
             }
